requirements.txt
# CLI TODO管理アプリ - 依存関係

# 自然言語処理機能に必要なパッケージ
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0

# Google Calendar API連携機能に必要なパッケージ
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pydantic>=1.10.0

# 画像生成機能に必要なパッケージ
torch>=2.0.0
diffusers>=0.21.0
# sdnqはGitHubからインストールが必要（Colabでは自動インストール）
# git+https://github.com/huggingface/diffusers
# git+https://github.com/Disty0/sdnq

# Python標準ライブラリのみ使用:
# - csv: CSVファイルの読み書き
# - os: ファイル存在確認
# - datetime: 日時処理
# - typing: 型ヒント

# 開発・テスト環境で使用する場合の推奨パッケージ（オプション）:
# pytest>=7.0.0          # テストフレームワーク
# black>=22.0.0           # コードフォーマッター
# flake8>=4.0.0          # リンター
# mypy>=0.950            # 型チェッカー